+++ microservices/tracksService.py
@@ -1,18 +1,11 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request
 from flask_restful import Resource, Api, marshal_with, fields
 from flask_sqlalchemy import SQLAlchemy
 from flask import abort
-
-import os
+from datetime import datetime
 
 tracksService = Flask(__name__)
 api = Api(tracksService)
-
-'''
-tracksService.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///tracksDB.db'
-tracksService.config['SECRET_KEY'] = 'secretkey'
-db = SQLAlchemy(tracksService)
-'''
 
 tracksService.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql+psycopg2://postgres:postgres@localhost:5432/music-browser'
 tracksService.config['SECRET_KEY'] = 'secretkey'
@@ -33,6 +26,7 @@ class Track(db.Model):
     title = db.Column(db.String, nullable=False)
     author = db.Column(db.String, nullable=False)
     track_url = db.Column(db.String, nullable=False)
+    createdAt = db.Column(db.String, nullable=False)
 
 
 # fields for JSON
@@ -40,7 +34,8 @@ trackFields = {
     'id': fields.Integer,
     'title': fields.String,
     'author': fields.String,
-    'track_url': fields.String
+    'track_url': fields.String,
+    'createdAt': fields.String
 }
 
 
@@ -63,12 +58,16 @@ class TracksEP(Resource):
         auth = headers.get("apiKey")
         if auth == '6327cc80-3093-4beb-90ee-191d69076366':
             data = request.json
-            track = Track(title=data['title'],
-                          author=data['author'],
-                          track_url=data['track_url'])
-            db.session.add(track)
-            db.session.commit()
-            return track
+            if ('title' in data) & ('author' in data) & ('track_url' in data):
+                track = Track(title=data['title'],
+                              author=data['author'],
+                              track_url=data['track_url'],
+                              createdAt=str(datetime.now())[:-7])
+                db.session.add(track)
+                db.session.commit()
+                return track
+            else:
+                return abort(400, 'Received JSON data is incorrect!')
         else:
             return abort(400, 'Provide correct api key!')
 
@@ -76,47 +75,59 @@ class TracksEP(Resource):
 class TrackEP(Resource):
     # get track with given id
     @marshal_with(trackFields)
-    def get(self, video_id):
+    def get(self, track_id):
         headers = request.headers
         auth = headers.get("apiKey")
         if auth == '6327cc80-3093-4beb-90ee-191d69076366':
-            track = Track.query.filter_by(id=video_id).first()
-            return track
+            track = Track.query.filter_by(id=track_id).first()
+            if track:
+                return track
+            else:
+                return abort(400, f'Track with id {track_id} does not exist!')
         else:
             return abort(400, 'Provide correct api key!')
 
     # update track with given id
     @marshal_with(trackFields)
-    def put(self, video_id):
+    def put(self, track_id):
         headers = request.headers
         auth = headers.get("apiKey")
         if auth == '6327cc80-3093-4beb-90ee-191d69076366':
             data = request.json
-            track = Track.query.filter_by(id=video_id).first()
-            track.title = data['title']
-            track.author = data['author']
-            track.track_url = data['track_url']
-            db.session.commit()
-            return track
+            if ('title' in data) & ('author' in data) & ('track_url' in data):
+                track = Track.query.filter_by(id=track_id).first()
+                if track:
+                    track.title = data['title']
+                    track.author = data['author']
+                    track.track_url = data['track_url']
+                    db.session.commit()
+                    return track
+                else:
+                    return abort(400, f'Track with id {track_id} does not exist!')
+            else:
+                return abort(400, 'Received JSON data is incorrect!')
         else:
             return abort(400, 'Provide correct api key!')
 
     # delete track with given id
     @marshal_with(trackFields)
-    def delete(self, video_id):
+    def delete(self, track_id):
         headers = request.headers
         auth = headers.get("apiKey")
         if auth == '6327cc80-3093-4beb-90ee-191d69076366':
-            track = Track.query.filter_by(id=video_id).first()
-            db.session.delete(track)
-            db.session.commit()
-            return track, 200
+            track = Track.query.filter_by(id=track_id).first()
+            if track:
+                db.session.delete(track)
+                db.session.commit()
+                return track
+            else:
+                return abort(400, f'Track with id {track_id} does not exist!')
         else:
             return abort(400, 'Provide correct api key!')
 
 
 api.add_resource(TracksEP, '/tracks')
-api.add_resource(TrackEP, '/track/<int:video_id>')
+api.add_resource(TrackEP, '/track/<int:track_id>')
 
 if __name__ == '__main__':
     tracksService.run(host="localhost", port=5000, debug=True)
